dc_cat/requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.2.0
chardet==4.0.0
discord.py==1.7.3
idna==3.4
multidict==6.0.4
numpy==1.21.6
pandas==1.3.5
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
typing_extensions==4.4.0
yarl==1.8.2
